+++ .gitignore
@@ -5,7 +5,6 @@
 /.idea/modules.xml
 /.idea/workspace.xml
 /.idea/caches
-.idea
 .DS_Store
 /build
 /captures
