+++ app/build.gradle
@@ -28,7 +28,5 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation 'com.android.support:design:28.0.0-rc02'
-    implementation 'commons-validator:commons-validator:1.6'
-    implementation 'com.google.android.gms:play-services-auth:15.0.1'
 }
 apply plugin: 'com.google.gms.google-services'
