+++ app/src/main/AndroidManifest.xml
@@ -9,16 +9,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Controllers.Welcome">
+        <activity android:name=".Welcome">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Controllers.Login.Login" />
-        <activity android:name=".Controllers.MainPage" />
-        <activity android:name=".Controllers.Register"></activity>
+        <activity android:name=".Login" />
+        <activity android:name=".MainPage" />
+        <activity android:name=".Register"></activity>
     </application>
 
 </manifest>
