--- app/src/main/java/edu/gatech/cs2340_68b/donationtracker/Controllers/Welcome.java
@@ -1,55 +1,0 @@
-package edu.gatech.cs2340_68b.donationtracker.Controllers;
-
-import android.content.Intent;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
-import android.graphics.drawable.Drawable;
-import android.support.constraint.ConstraintLayout;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageButton;
-
-import com.google.firebase.database.FirebaseDatabase;
-
-import edu.gatech.cs2340_68b.donationtracker.Controllers.Login.Login;
-import edu.gatech.cs2340_68b.donationtracker.R;
-
-public class Welcome extends AppCompatActivity {
-
-    private ImageButton imageButton;
-    private Button loginButton;
-    private Button registerButton;
-    private ActionBar actionBar;
-
-    private FirebaseDatabase database = FirebaseDatabase.getInstance();
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.welcome);
-
-        actionBar = getSupportActionBar();
-        actionBar.setBackgroundDrawable(new ColorDrawable(Color.parseColor("#1C2331")));
-        loginButton = (Button) findViewById(R.id.LoginButton);
-        registerButton = (Button) findViewById(R.id.RegisterButton);
-
-
-        loginButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(Welcome.this, Login.class);
-                startActivity(intent);
-            }
-        });
-
-        registerButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(Welcome.this, Register.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
