--- app/src/main/java/edu/gatech/cs2340_68b/donationtracker/ForgetPassword.java
@@ -1,31 +1,0 @@
-package edu.gatech.cs2340_68b.donationtracker;
-
-import android.os.Bundle;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.Button;
-
-
-
-public class ForgetPassword extends AppCompatActivity {
-    private ActionBar actionBar;
-    private Button logout;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.forget_password);
-        logout = (Button) findViewById(R.id.logoutButton);
-        logout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
-        actionBar = getSupportActionBar();
-        actionBar.setTitle("Reset Your Password");
-//        actionBar.setIcon(R.drawable.icon);
-//        actionBar.setDisplayUseLogoEnabled(true);
-    }
-}
