+++ app/src/main/java/edu/gatech/cs2340_68b/donationtracker/MainPage.java
@@ -1,20 +1,15 @@
-package edu.gatech.cs2340_68b.donationtracker.Controllers;
+package edu.gatech.cs2340_68b.donationtracker;
 
 import android.content.Intent;
-import android.graphics.drawable.ColorDrawable;
-import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-import edu.gatech.cs2340_68b.donationtracker.R;
-
 public class MainPage extends AppCompatActivity {
 
     private Button logout;
-    private ActionBar actionBar;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -27,10 +22,6 @@ public class MainPage extends AppCompatActivity {
                 finish();
             }
         });
-        actionBar = getSupportActionBar();
-//        actionBar.setTitle("");
-//        actionBar.setIcon(R.drawable.icon);
-//        actionBar.setDisplayUseLogoEnabled(true);
     }
 }
