+++ app/src/main/java/edu/gatech/cs2340_68b/donationtracker/User.java
@@ -1,8 +1,4 @@
-package edu.gatech.cs2340_68b.donationtracker.Models;
-
-import java.math.BigInteger;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
+package edu.gatech.cs2340_68b.donationtracker;
 
 public class User {
     private String username = "user";
@@ -66,6 +62,4 @@ public class User {
     public void setLastLogin(int lastLogin) {
         this.lastLogin = lastLogin;
     }
-
-
 }
