+++ app/src/main/res/layout/main_page.xml
@@ -4,8 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Controllers.MainPage"
-    android:background="#fff">
+    tools:context=".MainPage">
 
     <Button
         android:id="@+id/logoutButton"
@@ -17,6 +16,5 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-    />
-
+        />
 </android.support.constraint.ConstraintLayout>
