+++ app/src/main/res/layout/welcome.xml
@@ -4,9 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#fff"
-    android:padding="16dp"
-    tools:context=".Controllers.Welcome">
+    tools:context=".Welcome">
 
     <Button
         android:id="@+id/RegisterButton"
@@ -14,8 +12,10 @@
         android:layout_height="63dp"
         android:layout_marginEnd="8dp"
         android:layout_marginStart="4dp"
-        android:layout_marginTop="24dp"
-        android:background="@drawable/registerbutton"
+        android:layout_marginTop="20dp"
+        android:autoText="true"
+        android:background="#5d72e8"
+        android:text="Register"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.51"
         app:layout_constraintStart_toStartOf="parent"
@@ -25,28 +25,27 @@
         android:id="@+id/LoginButton"
         android:layout_width="172dp"
         android:layout_height="63dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginTop="8dp"
-        android:background="@drawable/loginbutton"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginEnd="8dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="68dp"
+        android:autoText="false"
+        android:background="#5d72e8"
+        android:elevation="0dp"
+        android:text="Login"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.631" />
+        app:layout_constraintTop_toBottomOf="@+id/imageButton" />
 
     <ImageButton
         android:id="@+id/imageButton"
-        android:layout_width="300dp"
-        android:layout_height="150dp"
-        android:layout_marginBottom="8dp"
+        android:layout_width="280dp"
+        android:layout_height="110dp"
         android:layout_marginEnd="8dp"
         android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginTop="68dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.185"
-        app:srcCompat="@drawable/donapp" />
+        app:srcCompat="@drawable/appd" />
 
 </android.support.constraint.ConstraintLayout>
